Moved all files of All-In-One tool (Tracker Alignment Validation) from 'Alignment/OfflineValidation/scripts' to 'Alignment/OfflineValidation/python/TkAlAllInOneTool' except the actual script to run the tool ('validateAlignments.py').
